(#16) refactor: csr 컴포넌트 분리
